requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==46.0.6
idna==3.11
pycparser==3.0
pycryptodome==3.23.0
python-dotenv==1.2.2
requests==2.33.1
tinytuya==1.17.6
urllib3==2.6.3
